moved cv2 preprocessing from helpers to a new file
